Split RulesetSettings into per-YAML forms (902 → 263 lines)
Moves RulesetSettings out of /settings/ (which is shared with the
user-facing settings modal) into /admin/sections/ where it belongs,
and splits the monolithic component by YAML file:

- RulesetSection.tsx          shell — state, fetch/save, file picker
- ruleset/meta.ts             FILE_META, DEFAULT_RULESETS, helpers
- ruleset/form-atoms.tsx      FormInput, FormTextarea, ToggleRow,
                              SectionButton
- ruleset/PersonalityForm.tsx personality.yaml
- ruleset/SecurityForm.tsx    security.yaml + permission tier editor
- ruleset/ParametersForm.tsx  parameters.yaml + agent param card
- ruleset/AccessForm.tsx      access.yaml + external-API editor

Each form file branches on the active section internally; the shell
just picks the right one based on the active YAML file. All files
under 265 lines.
