+++ client/src/components/admin/sections/ruleset/form-atoms.tsx
@@ -1,0 +1,116 @@
+import { Input } from "@/components/ui/input";
+import { Switch } from "@/components/ui/switch";
+import { Textarea } from "@/components/ui/textarea";
+
+export function FormInput({
+  label,
+  value,
+  onChange,
+  placeholder,
+  type = "text",
+}: {
+  label: string;
+  value: string | number;
+  onChange: (value: string) => void;
+  placeholder?: string;
+  type?: string;
+}) {
+  return (
+    <label className="space-y-2">
+      <div className="text-xs font-medium uppercase tracking-[0.18em] text-muted-foreground">
+        {label}
+      </div>
+      <Input
+        type={type}
+        value={String(value ?? "")}
+        onChange={(e) => onChange(e.target.value)}
+        placeholder={placeholder}
+        className="border-white/10 bg-black/30 text-sm"
+      />
+    </label>
+  );
+}
+
+export function FormTextarea({
+  label,
+  value,
+  onChange,
+  placeholder,
+  rows = 5,
+  hint,
+}: {
+  label: string;
+  value: string;
+  onChange: (value: string) => void;
+  placeholder?: string;
+  rows?: number;
+  hint?: string;
+}) {
+  return (
+    <label className="space-y-2">
+      <div className="text-xs font-medium uppercase tracking-[0.18em] text-muted-foreground">
+        {label}
+      </div>
+      <Textarea
+        rows={rows}
+        value={value}
+        onChange={(e) => onChange(e.target.value)}
+        placeholder={placeholder}
+        className="zed-glass border-white/10 text-sm"
+      />
+      {hint ? <p className="text-xs leading-5 text-muted-foreground">{hint}</p> : null}
+    </label>
+  );
+}
+
+export function ToggleRow({
+  label,
+  checked,
+  onCheckedChange,
+  description,
+}: {
+  label: string;
+  checked: boolean;
+  onCheckedChange: (checked: boolean) => void;
+  description: string;
+}) {
+  return (
+    <div className="flex items-center justify-between rounded-2xl border border-white/10 bg-black/20 px-4 py-3">
+      <div className="space-y-1 pr-4">
+        <div className="text-sm font-medium">{label}</div>
+        <div className="text-xs leading-5 text-muted-foreground">{description}</div>
+      </div>
+      <Switch checked={checked} onCheckedChange={onCheckedChange} />
+    </div>
+  );
+}
+
+export function SectionButton({
+  active,
+  label,
+  description,
+  onClick,
+}: {
+  active: boolean;
+  label: string;
+  description: string;
+  onClick: () => void;
+}) {
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+      className={[
+        "rounded-2xl border px-4 py-3 text-left transition-all",
+        active
+          ? "border-cyan-400/35 bg-cyan-400/10 shadow-[0_0_30px_rgba(34,211,238,0.08)]"
+          : "border-white/10 bg-black/25 hover:border-white/20 hover:bg-black/35",
+      ].join(" ")}
+    >
+      <div className="space-y-1">
+        <div className="text-sm font-medium">{label}</div>
+        <div className="text-xs leading-5 text-muted-foreground">{description}</div>
+      </div>
+    </button>
+  );
+}
